// src/services/models/response-result-recommend-danger-factor-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { RecommendDangerFactorResponse } from './recommend-danger-factor-response';

/**
 * 
 * @export
 * @interface ResponseResultRecommendDangerFactorResponse
 */
export interface ResponseResultRecommendDangerFactorResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultRecommendDangerFactorResponse
     */
    'code'?: string;
    /**
     * 
     * @type {RecommendDangerFactorResponse}
     * @memberof ResponseResultRecommendDangerFactorResponse
     */
    'data'?: RecommendDangerFactorResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultRecommendDangerFactorResponse
     */
    'message'?: string;
}