// src/services/models/response-result-get-company-danger-factor-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { GetCompanyDangerFactorResponse } from './get-company-danger-factor-response';

/**
 * 
 * @export
 * @interface ResponseResultGetCompanyDangerFactorResponse
 */
export interface ResponseResultGetCompanyDangerFactorResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetCompanyDangerFactorResponse
     */
    'code'?: string;
    /**
     * 
     * @type {GetCompanyDangerFactorResponse}
     * @memberof ResponseResultGetCompanyDangerFactorResponse
     */
    'data'?: GetCompanyDangerFactorResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetCompanyDangerFactorResponse
     */
    'message'?: string;
}