// src/services/models/response-result-get-company-process-title-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { GetCompanyProcessTitleResponse } from './get-company-process-title-response';

/**
 * 
 * @export
 * @interface ResponseResultGetCompanyProcessTitleResponse
 */
export interface ResponseResultGetCompanyProcessTitleResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetCompanyProcessTitleResponse
     */
    'code'?: string;
    /**
     * 
     * @type {GetCompanyProcessTitleResponse}
     * @memberof ResponseResultGetCompanyProcessTitleResponse
     */
    'data'?: GetCompanyProcessTitleResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetCompanyProcessTitleResponse
     */
    'message'?: string;
}