// src/services/models/response-result-get-danger-factor-title-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { GetDangerFactorTitleResponse } from './get-danger-factor-title-response';

/**
 * 
 * @export
 * @interface ResponseResultGetDangerFactorTitleResponse
 */
export interface ResponseResultGetDangerFactorTitleResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetDangerFactorTitleResponse
     */
    'code'?: string;
    /**
     * 
     * @type {GetDangerFactorTitleResponse}
     * @memberof ResponseResultGetDangerFactorTitleResponse
     */
    'data'?: GetDangerFactorTitleResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultGetDangerFactorTitleResponse
     */
    'message'?: string;
}