// src/services/models/response-result-upsert-company-danger-factor-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { UpsertCompanyDangerFactorResponse } from './upsert-company-danger-factor-response';

/**
 * 
 * @export
 * @interface ResponseResultUpsertCompanyDangerFactorResponse
 */
export interface ResponseResultUpsertCompanyDangerFactorResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultUpsertCompanyDangerFactorResponse
     */
    'code'?: string;
    /**
     * 
     * @type {UpsertCompanyDangerFactorResponse}
     * @memberof ResponseResultUpsertCompanyDangerFactorResponse
     */
    'data'?: UpsertCompanyDangerFactorResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultUpsertCompanyDangerFactorResponse
     */
    'message'?: string;
}